Extract concern into its own file
